Commands sorted in Command file + Makefile refait dossier - sous dossier crees
