+++ srcs/utils.cpp
@@ -1,4 +1,4 @@
-#include "Utils.hpp"
+#include "utils.hpp"
 #include <unistd.h>
 #include <signal.h>
